pyproject.toml
[project]
name = "voiceeval-sdk"
version = "0.1.0"
description = "Enterprise-grade Observability and Evaluation SDK for Voice Agents"
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [
    { name = "VoiceEval Team", email = "hello@voiceeval.com" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
urls.Repository = "https://github.com/voiceeval/voiceeval-sdk"
urls.Homepage = "https://voiceeval.com"
dependencies = [
    "fastapi>=0.127.1",
    "httpx>=0.28.1",
    "openai>=2.14.0",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-exporter-otlp>=1.39.1",
    "opentelemetry-instrumentation-anthropic>=0.50.1",
    "opentelemetry-instrumentation-google-generativeai>=0.50.1",
    "opentelemetry-instrumentation-openai>=0.50.1",
    "opentelemetry-sdk>=1.39.1",
    "pydantic>=2.0.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/voiceeval"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
